// src/utils/config/env.enum.ts
export enum ENV {
  NODE_ENV = 'NODE_ENV',
  DATABASE_URL = 'DATABASE_URL',
  DIRECT_URL = 'DIRECT_URL',
  PORT = 'PORT',
  JWT_SECRET = 'JWT_SECRET',
  ADMIN_TOKEN = 'ADMIN_TOKEN',

  BREVO_API_KEY = 'BREVO_API_KEY',
  BREVO_SENDER_EMAIL = 'BREVO_SENDER_EMAIL',

  COMPANY_NAME = 'COMPANY_NAME',
  COMPANY_EMAIL = 'COMPANY_EMAIL',

  REDIS_URL = 'REDIS_URL',
  REDIS_TTL = 'REDIS_TTL',
  CACHE_MAX = 'CACHE_MAX',
  CACHE_TTL = 'CACHE_TTL',

  STORAGE_PROVIDER = 'STORAGE_PROVIDER',
  MAX_FILE_SIZE = 'MAX_FILE_SIZE',
  AWS_REGION = 'AWS_REGION',
  AWS_ENDPOINT = 'AWS_ENDPOINT',
  S3_PUBLIC_BUCKET_NAME = 'S3_PUBLIC_BUCKET_NAME',
  S3_PRIVATE_BUCKET_NAME = 'S3_PRIVATE_BUCKET_NAME',
  AWS_ACCESS_KEY_ID = 'AWS_ACCESS_KEY_ID',
  AWS_SECRET_ACCESS_KEY = 'AWS_SECRET_ACCESS_KEY',

  FRONTEND_URL = 'FRONTEND_URL',
}